perf: increase frequency of reindexing job
Index being stored as cache means it will be recalculated when it
expired. It should be safe to frequently trigger reindexing schedule
since it won't actually doing reindex when the cache is warm.

The change in cache ttl is meant to make sure that the new content will
be available faster.